+++ src/test/mocks/SyncablePool.sol
@@ -9,7 +9,7 @@ contract SyncablePool is Pool, ISyncablePool {
     constructor(
         address shares_,
         address fyToken_,
-        uint128 ts_,
+        uint256 ts_,
         uint16 g1Fee_
     ) Pool(shares_, fyToken_, ts_, g1Fee_) {}
